+++ tests/index.spec.js
@@ -2,11 +2,6 @@ const fs = require('fs');
 const path = require('path');
 const expect = require('chai').expect;
 
-const {
-  execute,
-  executeRequireStatement
-} = require('../src/execute.js');
-
 const abellRenderer = require('../src/index.js');
 
 describe('render() - renders abellTemplate into HTML Text', () => {
@@ -33,7 +28,10 @@ describe('render() - renders abellTemplate into HTML Text', () => {
         .render(
           abellTemplate, 
           sampleSandbox, 
-          {basePath: path.join(__dirname, 'resources')}
+          {
+            basePath: path.join(__dirname, 'resources'), 
+            allowRequire: true
+          }
         )
     ).to.equal(htmlTemplate);
   });
@@ -79,9 +77,28 @@ describe('render() - renders abellTemplate into HTML Text', () => {
     expect(
       abellRenderer.render(
         abellTemplate,
-        {}
+        {},
+        {allowRequire: true}
       ).trim()
     ).to.equal('<div>8 hi/hello hi/hello</div>');
+  });
+
+  it('should throw an error if require() is used without allowRequire: true option', () => {
+    const abellTemplate = `
+      {{
+        const path = require('path');
+        const hiHelloPath = require('path').join('hi', 'hello');
+      }}
+      <div>{{ path.join('hi', 'hello') }} {{ hiHelloPath }}</div>
+    `;
+
+    expect(
+      () => 
+        abellRenderer.render(
+          abellTemplate,
+          {},
+        )
+    ).to.throw('require() is not allowed in the script');
   });
 
   it('should not throw error and return same value if blank brackets passed', () => {
@@ -100,43 +117,5 @@ describe('render() - renders abellTemplate into HTML Text', () => {
         {}
       )
     ).to.equal('{{ This is ignored }}');
-  });
-});
-
-
-describe('execute() - Executes JavaScript passed to it as string', () => {
-  it('should output added value when addition is performed on two values', () => {
-    expect(
-      execute('24 + 12', {}).value
-    ).to.equal(36);
-  });
-
-  it('should update value of a to new value', () => {
-    expect(
-      execute('a = 22 + 22', {a: 4}).sandbox.a
-    ).to.equal(44);
-  });
-
-  it('should not update value that is inside string', () => {
-    expect(
-      execute('(() => \'a = b\')()').value
-    ).to.equal('a = b');
-  });
-});
-
-
-describe('executeRequireStatement() - executes the code with require() in its string', () => {
-  it('should add path native object when required', () => {
-    expect(
-      executeRequireStatement('const path = require(\'path\')')
-        .path.join('test', 'path')
-    ).to.equal(path.join('test', 'path'));
-  });
-
-  it('should handle the case of require(\'module\').property', () => {
-    expect(
-      executeRequireStatement('const testPath = require(\'path\').join(\'test\',\'path\')')
-        .testPath
-    ).to.equal(path.join('test', 'path'));
   });
 });